+++ pygameMenu/widgets/highlight/__init__.py
@@ -3,8 +3,8 @@
 pygame-menu
 https://github.com/ppizarror/pygame-menu
 
-LABEL
-Label class, adds a simple text to the Menu.
+HIGHLIGHT
+This module contains the widget highlight effects.
 
 License:
 -------------------------------------------------------------------------------
@@ -29,44 +29,3 @@ WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
 CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 -------------------------------------------------------------------------------
 """
-
-from pygameMenu.widgets.widget import Widget
-
-
-class Label(Widget):
-    """
-    Label widget.
-
-    :param label: Text of the button
-    :type label: basestring
-    :param label_id: Button ID
-    :type label_id: basestring
-    """
-
-    def __init__(self, label, label_id=''):
-        assert isinstance(label, str)
-        super(Label, self).__init__(widget_id=label_id)
-        self._label = label
-        self.is_selectable = False
-
-    def _apply_font(self):
-        pass
-
-    # noinspection PyMissingOrEmptyDocstring
-    def draw(self, surface):
-        self._render()
-        surface.blit(self._surface, self._rect.topleft)
-
-    def _render(self):
-        if not self._render_hash_changed(self._label, self._font_color):
-            return
-        self._surface = self._render_string(self._label, self._font_color)
-        self._rect.width, self._rect.height = self._surface.get_size()
-
-    # noinspection PyMissingOrEmptyDocstring
-    def draw_selected_rect(self, surface, selected_color, inflatex, inflatey, border_width):
-        pass  # Nothing to select
-
-    # noinspection PyMissingOrEmptyDocstring
-    def update(self, events):
-        return False
