+++ pygameMenu/widgets/widget/vmargin.py
@@ -3,8 +3,8 @@
 pygame-menu
 https://github.com/ppizarror/pygame-menu
 
-WIDGETS
-Widgets elements that can be added to the menu.
+VERTICAL MARGIN
+Vertical box margin.
 
 License:
 -------------------------------------------------------------------------------
@@ -31,13 +31,34 @@ CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 """
 
 from pygameMenu.widgets.core.widget import Widget
+from pygameMenu.utils import make_surface
 
-from pygameMenu.widgets.widget.button import Button
-from pygameMenu.widgets.widget.colorinput import ColorInput
-from pygameMenu.widgets.widget.image import Image
-from pygameMenu.widgets.widget.label import Label
-from pygameMenu.widgets.widget.menubar import MenuBar
-from pygameMenu.widgets.widget.scrollbar import ScrollBar
-from pygameMenu.widgets.widget.selector import Selector
-from pygameMenu.widgets.widget.textinput import TextInput
-from pygameMenu.widgets.widget.vmargin import VMargin
+
+class VMargin(Widget):
+    """
+    Vertical margin widget.
+    """
+
+    def __init__(self):
+        super(VMargin, self).__init__()
+        self.is_selectable = False
+
+    def _apply_font(self):
+        self._font_size = 0
+        self._shadow = False
+
+    # noinspection PyMissingOrEmptyDocstring
+    def draw(self, surface):
+        self._render()
+        surface.blit(self._surface, self._rect.topleft)
+
+    def _render(self):
+        if self._surface is not None:
+            return
+        self._surface = make_surface(1, 1, alpha=True)
+        self._rect.width = 0.0
+        self._rect.height = 0.0
+
+    # noinspection PyMissingOrEmptyDocstring
+    def update(self, events):
+        return False
